Add DELETE_EVENT_REF for `delete` events
When a delete event occurs, the default value for the `GITHUB_REF`
passed to the action is the default branch (https://developer.github.com/actions/managing-workflows/workflow-configuration-options/#events-supported-in-workflow-files). It might be useful to know what was deleted when triggering an action off of a delete event. This change will pass the `ref` of the deleted branch/tag as a custom `env` variable on the created build